Big refactoring to make the data structures have better names and a structure that is easier to work with.
Also, it is now an error if two attributes named "version" are found
in a set.

Some details and explanation:

Moved all the data structures to Types.hs so that we can just look in
one place to understand the structure of these trees that we are
using.  This is similar to how hnix is structured.

Renamed UpdateSource to FetchTree.  It's a tree of information about
fetches; that's the main thing in it.  Yes, it helps us performs
updates, but we already knew that from the namespace it is in
(Update.Nix.FetchGit).

Made it so we can pass around a single FetchTree instead of an array
of them.  This means we can delete traverse2, sequenceA2, and
mapConcurrently2.  This also makes FetchTree feel more like NExprLoc.

The new structure of FetchTree makes it easy add other updatable
attributes in the future (updating "name" would be useful for many
packages as they are written today).  Also, by supplying something
else for its type parameter, we can expand it to handle non-git
fetches like fetchurl.

Renamed some functions to be less confusing and match the new data
structure names.

Reordered the cases in extractAttr so that errors are grouped
together.

The code is 16 lines shorter now.