Major changes to the order of operations.
Separated two processes: the process of fetching the latest info from
the internet and the process of deciding what spans in the file to
update.  I think this is needed for later when we update the version
attributes, or at least it will make that code easier to understand.

Linearization of source locations is now done at the end, right before
we need it.  This means that we have nice column/line source locations
throughout most parts of the code, which we can use to produce
friendly error messages.  This linearization was one of the
motivations for having the "Massaging the data a little" phase and the
FetchGitUpdateInfo type; those can be removed now, making the program
simpler.

Simplified updateSpans by applying the updates in reverse order
instead of adjusting each update to account for the updates the come
before it.

Changed the names for the SpanUpdate members to hopefully be more
clear.