Extract and parse commit date information from the output of nix-prefetch-git.
We are no longer calling nix-shell: dependency installation should be
done at the time that this program is installed and it should be
handled by whatever package manager the user chooses to use.  We can
just assume that the user has nix-prefetch-git on their PATH.

Also, in my environment, calling nix-shell meant that we would end up
using the version of nix-prefetch-git from the unstable channel, which
does not have the latest changes in it to output the date.