Remove the ArgNotASet error.
I think Nix function applications always use a set.  It is cleaner to
just ignore the contrary case then to add an error code for it and add
code to describe that error to the user.

We can add the error code back later if it turns out to be important.